blog_api/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
factory-boy==3.2.1
Faker==15.3.1
Pillow==9.3.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
